requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-telegram-bot>=20.7
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
psutil>=5.9.0
